server/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# CORS & Middleware
python-dotenv==1.0.0

# Payment Integration (for escrow system)
stripe==7.11.0

# File Upload & Storage
python-magic==0.4.27
aiofiles==23.2.1

# Email & Notifications
fastapi-mail==1.4.1

# Date & Time
python-dateutil==2.8.2

# Validation & Utils
email-validator==2.1.0
phonenumbers==8.13.27

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# API Documentation
fastapi-pagination==0.12.14

# Caching (optional but useful)
redis==5.0.1

# Image Processing (for vendor portfolios, wedding photos)
pillow==10.2.0

# Background Tasks
celery==5.3.6
